Close #507: [`refined4s-extras-render`] Reorganize type-class instances in the modules
